tool vs cli distinction
